+++ src/beliefprop/BpOptimizeCPU/RunBpOnStereoSetOptimizedCPU.h
@@ -74,7 +74,7 @@ inline std::optional<beliefprop::BpRunOutput> RunBpOnStereoSetOptimizedCPU<T, DI
   //set number of threads to use when running code in parallel using OpenMP from input parallel parameters
   //current setting on CPU is to execute all parallel processing in a run using the same number of parallel threads
   const unsigned int nthreads = 
-    parallel_params.OptParamsForKernel({static_cast<unsigned int>(beliefprop::BpKernel::kBlurImages), 0})[0];
+    parallel_params.OptParamsForKernel({static_cast<size_t>(beliefprop::BpKernel::kBlurImages), 0})[0];
   #ifndef __APPLE__
     omp_set_num_threads(nthreads);
   #endif //__APPLE__
